requirements.txt
streamlit
matplotlib
japanize-matplotlib
numpy
Pillow
